Rename defaults.ts to index.ts
